Split US to match new PSP file
